requirements.txt
django==5.0.2
django-allauth==0.61.1
PyGithub==2.2.0
langchain_openai==0.0.7
langchain==0.1.9
langchain-community==0.0.24
langchainhub==0.1.14
openai==1.10.0
GitPython==3.1.42
instructor==0.3.5